cic(#1396): move the fan-out pair out of the switch, at zero new fields
`/ame` and `/amsg` were the last pair anyone believed the send pipeline
held back. It does not, and the belief outlived its cause: every value
the arm reads is an ordinary module import — `sendFanOut`
(`./sendPipeline`), `requestConfirm` (`./confirmDialog`),
`joinedChannelsOnNetwork` (`./joinedChannels`), `friendlyError`. It
touches the composer buffer nowhere.

Two lifts made that true, both under #1513 and neither of them about
this pair: `6ef878fe` moved `sendFanOut` out of `compose.ts`, where it
was a module-level const that `commands/` could not import because it
was never exported, and `5779c755` lifted `joinedChannelsOnNetwork` out
of the controller closure, which is the one that had really been a
closure. Nothing about the pair changed; the ground under it did.

So the record gains no field. `FANOUT_CONFIRM_THRESHOLD` travels with
the handler — it had exactly one reader and the move keeps it that way,
rather than leaving a const in `compose.ts` for a body that no longer
lives there.

ONE handler for two kinds (`CommandHandler<"ame" | "amsg">`): the pair
differ only by the ACTION framing that `cmd.kind` already decides, and
splitting them would duplicate the confirm gate and the partial-failure
accounting.

The arm's two error paths were early `return`s inside the switch and are
now the handler's returns, assigned to `result`. Equivalent: nothing
sits between the `break` and the `return result` that could read or
rewrite it.
