+++ cicchetto/src/lib/compose.ts
@@ -5,6 +5,7 @@ import { token } from "./auth";
 import { type ChannelKey, channelKey, decodeChannelKey } from "./channelKey";
 import { isChannelName } from "./chantypes";
 import type { CommandContext, DispatchOutcome, SubmitResult } from "./commands/context";
+import { fanOutCommand } from "./commands/fanout";
 import { watchlistCommand } from "./commands/highlight";
 import {
   aliasDefineCommand,
@@ -62,7 +63,6 @@ import {
 } from "./commands/session";
 import { topicClearCommand, topicSetCommand, topicShowCommand } from "./commands/topic";
 import { joinCommand, listCommand, partCommand, queryCommand } from "./commands/window";
-import { requestConfirm } from "./confirmDialog";
 import { documentTeardownEpoch, documentTornDownSince } from "./documentTeardown";
 import { type FramePreview, framePreview } from "./frameBudget";
 import { friendlyError } from "./friendlyError";
@@ -79,7 +79,7 @@ import { canonicalQueryNick, openQueryWindowState } from "./queryWindows";
 // different recipient while echoing here, so it is named for the window, not
 // for one of the verbs it can carry.
 import { selectedChannel, setSelectedChannel } from "./selection";
-import { draftLines, sendBodyLines, sendFanOut, wireBody } from "./sendPipeline";
+import { draftLines, sendBodyLines, wireBody } from "./sendPipeline";
 import { isServicesSender } from "./servicesSender";
 import { parseSlash } from "./slashCommands";
 import { SERVER_WINDOW_NAME } from "./windowKinds";
@@ -208,12 +208,6 @@ const persistDrafts = (states: Record<ChannelKey, ComposeState>): void => {
   if (keys.length === 0) sessionStorage.removeItem(DRAFTS_KEY);
   else sessionStorage.setItem(DRAFTS_KEY, JSON.stringify(drafts));
 };
-
-// #431 — above this many channels an /ame|/amsg asks before it writes anything.
-// A DECIDED number (vjt, 2026-08-09), not a tuned one: one keystroke that
-// writes to N channels is a spam vector, and ten is where a fan-out stops
-// reading as "the rooms I am in" and starts reading as a broadcast.
-const FANOUT_CONFIRM_THRESHOLD = 10;
 
 /**
  * #1108 — what the current draft will do to the wire, or `null` when there is
@@ -784,68 +778,9 @@ const exports_ = identityScopedStore((onIdentityChange) => {
           result = r;
           break;
         }
-        // #431 — /ame + /amsg: the same body to EVERY joined channel of THIS
-        // network. Deliberately not addressed to the active window: the fan-out
-        // is per-NETWORK (vjt 2026-08-09 — cross-network is out of scope, and
-        // no flag for it), so it works from the $server window just as well as
-        // from a channel, and it includes the window it was typed in.
         case "ame":
         case "amsg": {
-          const action = cmd.kind === "ame";
-          // The joined-channel projection #30's channel tab-completion already
-          // reads: server-owned, so `invited` (the window an unsolicited INVITE
-          // opens), `pending`, `failed`, `kicked` and `parked` are all excluded
-          // by the same `joined` predicate, and there is no parallel
-          // client-side channel list to drift from it. Queries and $server
-          // cannot appear at all — `window_states` is channel-keyed by
-          // construction on the server, a DM lives in `queryWindows`.
-          const targets = joinedChannelsOnNetwork(key).sort();
-          if (targets.length === 0) {
-            return { error: `/${cmd.kind}: no joined channel on ${networkSlug}` };
-          }
-          // Channels the fan-out has finished, for the "how far did it get"
-          // half of a failure — the number that tells the operator whether
-          // retyping the line would double-send it.
-          let done = 0;
-          const run = (): Promise<void> =>
-            sendFanOut(networkSlug, targets, cmd.body, action, (n) => {
-              done = n;
-            });
-          if (targets.length > FANOUT_CONFIRM_THRESHOLD) {
-            // The question names every target, not just the count: the blast
-            // radius IS the thing being consented to, so "11 channels" alone
-            // would be consent without disclosure.
-            requestConfirm({
-              title: action ? "Send action to every channel" : "Send message to every channel",
-              body: `Send to ${targets.length} channels on ${networkSlug}? ${targets.join(", ")}`,
-              confirmLabel: "Send",
-              // Detached: `requestConfirm` resolves nothing, so the send cannot
-              // be awaited here and its failure cannot surface inline the way
-              // the un-gated path's does below. Logged with a grep key instead,
-              // mirroring `windowClose.disconnectNetwork` — the same shape of
-              // destructive-action-behind-a-confirm. Never bare: an unhandled
-              // rejection here would be a fan-out that stopped in silence.
-              onConfirm: () => {
-                void run().catch((err) => {
-                  console.warn(
-                    `[/${cmd.kind}] fan-out stopped after ${done} of ${targets.length} channels on ${networkSlug}:`,
-                    err,
-                  );
-                });
-              },
-              alternative: null,
-            });
-            result = { ok: `/${cmd.kind}: ${targets.length} channels — confirm to send` };
-            break;
-          }
-          try {
-            await run();
-          } catch (e) {
-            return {
-              error: `${friendlyError(e)} — sent to ${done} of ${targets.length} channels`,
-            };
-          }
-          result = { ok: `/${cmd.kind}: sent to ${targets.join(", ")}` };
+          result = await fanOutCommand(cmd, ctx);
           break;
         }
         case "ctcp": {
